Fixed error handleing approch

+++ 7-api-with-go/internal/domain/notes/domain/repository.go
@@ -4,6 +4,6 @@ import "context"
 
 type Repository interface {
 	FindAll(ctx context.Context) ([]*Note, error)
-	FindById(ctx context.Context, id int) (*Note, error)
+	FindById(ctx context.Context, id int64) (*Note, error)
 	Create(ctx context.Context, note *Note) (*Note, error)
 }
